Types refactor
