+++ exporter/SynthesisFusionAddin/src/Parser/SynthesisParser/Joints.py
@@ -33,7 +33,8 @@ from proto.proto_out import assembly_pb2, joint_pb2, motor_pb2, signal_pb2, type
 
 from ...general_imports import *
 from ...Logging import getLogger
-from ..ExporterOptions import ExporterOptions, JointParentType, SignalType
+from ...Types import JointParentType, SignalType
+from ..ExporterOptions import ExporterOptions
 from .PDMessage import PDMessage
 from .Utilities import construct_info, fill_info, guid_occurrence
