+++ exporter/SynthesisFusionAddin/src/Parser/SynthesisParser/Parser.py
@@ -10,8 +10,9 @@ from proto.proto_out import assembly_pb2, types_pb2
 from ...APS.APS import upload_mirabuf  # This line causes everything to break
 from ...general_imports import *
 from ...Logging import getLogger, logFailure, timed
+from ...Types import ExportLocation, ExportMode
 from ...UI.Camera import captureThumbnail, clearIconCache
-from ..ExporterOptions import ExporterOptions, ExportLocation, ExportMode
+from ..ExporterOptions import ExporterOptions
 from . import Components, JointHierarchy, Joints, Materials, PDMessage
 from .Utilities import *
