+++ exporter/SynthesisFusionAddin/src/UI/Configuration/SerialCommand.py
@@ -7,7 +7,7 @@ This is used to store in the userdata.
 
 import json
 
-from ...Types.OString import OString
+from ...Types import OString
 
 
 def generateFilePath() -> str:
